[RDRViews] restructuring
