+++ src/semantic_world/views/world_rdr/world_rdr.py
@@ -2,7 +2,7 @@ from typing_extensions import Dict, Any
 from ripple_down_rules.rdr import GeneralRDR
 from ripple_down_rules.datastructures.case import Case, create_case
 from semantic_world.world import World
-from . import world_views_mcrdr_of_grdr as views_classifier
+from . import world_views_mcrdr as views_classifier
 
 
 classifiers_dict = dict()
