+++ test/test_views/test_views.py
@@ -18,7 +18,7 @@ from ripple_down_rules.experts import Human
 from ripple_down_rules.rdr import GeneralRDR
 from semantic_world.adapters.urdf import URDFParser
 from semantic_world.views import *
-from semantic_world.views.view_classifiers import world_rdr
+from semantic_world.views.world_rdr import world_rdr
 from semantic_world.world import World
 
 
@@ -42,9 +42,8 @@ class ViewTestCase(unittest.TestCase):
     kitchen = os.path.join(urdf_dir, "kitchen-small.urdf")
     apartment = os.path.join(urdf_dir, "apartment.urdf")
     main_dir = os.path.join(os.path.dirname(__file__), '..', '..')
-    views_dir = os.path.join(main_dir, "src/semantic_world/views/view_classifiers")
-    views_json_dir = os.path.join(main_dir, "src/semantic_world/views/rdr_json")
-    views_rdr_filename = os.path.join(views_json_dir, "views_rdr")
+    views_dir = os.path.join(main_dir, "src/semantic_world/views")
+    views_rdr_model_name = "world_rdr"
     test_dir = os.path.join(main_dir, 'tests')
     expert_answers_dir = os.path.join(test_dir, "test_expert_answers")
     app: Optional[QApplication] = None
@@ -57,7 +56,7 @@ class ViewTestCase(unittest.TestCase):
         cls.apartment_parser = URDFParser(cls.apartment)
         if RDRCaseViewer is not None and QApplication is not None and cls.use_gui:
             cls.app = QApplication(sys.argv)
-            cls.viewer = RDRCaseViewer(save_file=cls.views_json_dir)
+            cls.viewer = RDRCaseViewer(save_file=cls.views_dir)
 
     def test_id(self):
         v1 = Handle(1)
@@ -103,20 +102,18 @@ class ViewTestCase(unittest.TestCase):
 
     def fit_views_and_get_rdr(self, world: World, required_views: List[Type[View]], save_rules: bool = True,
                               use_current_rules: bool = True, update_existing_views: bool = False) -> GeneralRDR:
-        expert = Human(viewer=self.viewer)
+        expert = Human(viewer=self.viewer, answers_save_path=os.path.join(self.views_dir, "expert_answers"))
         if use_current_rules:
-            grdr = GeneralRDR.load(self.views_rdr_filename)
-            grdr.update_from_python_file(self.views_dir)
+            grdr = GeneralRDR.load(self.views_dir, self.views_rdr_model_name)
             grdr.set_viewer(self.viewer)
         else:
-            grdr = GeneralRDR(viewer=self.viewer)
+            grdr = GeneralRDR(save_dir=self.views_dir, viewer=self.viewer)
 
         for view in required_views:
             case_query = CaseQuery(world, "views", (view,), False)
             if update_existing_views or case_query.attribute_name not in grdr.start_rules_dict:
                 grdr.fit_case(CaseQuery(world, "views", (view,), False), expert=expert)
         if save_rules:
-            grdr.write_to_python_file(self.views_dir)
-            grdr.save(self.views_rdr_filename)
+            grdr.save(self.views_dir, self.views_rdr_model_name)
 
         return grdr
